+++ packages/zowe-explorer/__tests__/__e2e__/step_definitions/dialogs/IssueCommand.steps.ts
@@ -20,22 +20,12 @@ Then("a quick pick appears to select a profile", async function () {
 });
 When("a user selects a profile", async function () {
     await expect(this.input).toBeDefined();
-    let qpItems = await this.input.getQuickPicks();
+    const qpItems = await this.input.getQuickPicks();
     await qpItems.at(0).select();
     await expect(this.input.elem).toBeDisplayedInViewport();
-
-    // Issue TSO command has an extra quick pick to select from
-    if (this.openedCommand === "Issue TSO Command") {
-        qpItems = await this.input.getQuickPicks();
-        await qpItems.at(0).select();
-    }
 });
 
 Then(/a user can enter in (.*) as the command and submit it/, async function (command: string) {
-    if (this.openedCommand === "Issue TSO Command") {
-        await this.input.setText("IZUACCT");
-        await this.input.confirm();
-    }
     await this.input.setText(command);
     await this.input.confirm();
 });